+++ Options.xaml.cs
@@ -13,6 +13,7 @@
 using System.ComponentModel.Composition;
 using System.Diagnostics;
 using System.Windows;
+using System.Windows.Controls;
 using System.Windows.Navigation;
 
 namespace DaleGhent.NINA.GroundStation {
@@ -27,6 +28,38 @@ namespace DaleGhent.NINA.GroundStation {
         private void Hyperlink_RequestNavigate(object sender, RequestNavigateEventArgs e) {
             _ = Process.Start(new ProcessStartInfo(e.Uri.AbsoluteUri));
             e.Handled = true;
+        }
+
+        private void PasswordBox_Smtp_PasswordChanged(object sender, RoutedEventArgs e) {
+            if (sender is PasswordBox elem) {
+                if (elem.DataContext is GroundStation vm) {
+                    vm.SetSmtpPassword(elem.SecurePassword);
+                }
+            }
+        }
+
+        private void PasswordBox_Smtp_Loaded(object sender, RoutedEventArgs e) {
+            if (sender is PasswordBox elem) {
+                if (elem.DataContext is GroundStation vm) {
+                    elem.Password = vm.SmtpPassword;
+                }
+            }
+        }
+
+        private void PasswordBox_Mqtt_PasswordChanged(object sender, RoutedEventArgs e) {
+            if (sender is PasswordBox elem) {
+                if (elem.DataContext is GroundStation vm) {
+                    vm.SetMqttPassword(elem.SecurePassword);
+                }
+            }
+        }
+
+        private void PasswordBox_Mqtt_Loaded(object sender, RoutedEventArgs e) {
+            if (sender is PasswordBox elem) {
+                if (elem.DataContext is GroundStation vm) {
+                    elem.Password = vm.MqttPassword;
+                }
+            }
         }
     }
 }